+++ 0x1E-search_algorithms/search_algos.h
@@ -5,7 +5,7 @@
 #include <math.h>
 int linear_search(int *array, size_t size, int value);
 int binary_search(int *array, size_t size, int value);
-void print_arr(int *array, size_t size, int left);
+void print_arr(int *array, size_t size, size_t left);
 int jump_search(int *array, size_t size, int value);
 int interpolation_search(int *array, size_t size, int value);
